Add mockData + tests
